refactored upcomingevents into separate component
