+++ client/components/index.js
@@ -12,3 +12,4 @@ export {PieChartData} from './PieChartData'
 export {default as InputPollForm} from './InputPollForm'
 export {default as Home} from './Home'
 export {default as UserDashboard} from './UserDashboard'
+export {default as UpcomingEvents} from './UpcomingEvents'
