+++ client/routes.js
@@ -39,7 +39,7 @@ class Routes extends Component {
         {isLoggedIn && (
           <Switch>
             {/* Routes placed here are only available after logging in */}
-            <Route path="/home" component={UserHome} />
+            <Route path="/home" component={UserDashboard} />
           </Switch>
         )}
         {/* Displays our Login component as a fallback */}
